Fold stack-map calls into the instruction they track
Calls to the llvm.experimental.stackmap intrinsic happen before
conditional branches, and after regular function calls. This stack-map
call encodes a 'safepoint' -- a record of all the live variables so that
these instructions can safely deoptimise if a guard fails.

Conceptually, safepoints 'belong' to these instructions, and having to
remember whether to go forward or back to find them is just LLVM
complexity which doesn't fit with our mental model how AOT IR should
look.

This commit refactors this and makes safepoints easier to work with. In
essence it does two thing:
    1) Removes the `DeoptSafepoint` instruction in the AOT IR, folding
       safepoint info into the relevant `CondBr` and `Call`
       instructions.
    2) Renames uses of stack-map (an implementation detail) in the AOT
       IR to safepoint.